Added support for Microsoft Visual C++ Compiler.
1) shz_cdefs.h
    a. compiler detection + newly defined macros
    b. #ifdefs for MSVC-specific implementations of GNUC
    directives/attributes (where applicable).
2) SHZ_INIT() introduced for generic initialization of temporary objects
   using C99 compound literals or C++20 designated initializers.
3) API-level C constexpr parameter checks which forward to GCC
   __builtin_xxx() routines for compile-time evaluation now #ifdef'd out
   for MSVC.
4) C++23 language detection needed to look for _MSC_LANG version, cuz
   MSVC is stupid.
5) SHZ_MEMORY_BARRIER_HARD/SOFT() moved to shz_cdefs and is not compiler-specific.
6) CMakeLists.txt
    a. Had to add find_package(cglm) to get fucking MSVC to find
    dependency for unit tests.
    b. Had to change C standard version for unit tests from C23 to C17
    cuz again, MSVC sucks.
7) Synced libGimbal submodule with some MSVC-specific fixes.

*** EVERYTHING BUILDING FINE + ALL UNIT TESTS PASSING WITH MSVC ***

+++ source/shz_complex.c
@@ -21,14 +21,14 @@ void shz_fft(shz_complex_t* s, size_t size) {
     const float a = -2.0f * SHZ_F_PI / (float)size; // Twiddle factor angle increments
 
     // This loop determines the stage of the FFT
-    for(int len = size >> 1; len > 0 ; len >>= 1) { 
+    for(size_t len = size >> 1; len > 0 ; len >>= 1) {
 
         // This loop determines which Twiddle factor to use
-        for(int j = 0; j < len; ++j) {
+        for(size_t j = 0; j < len; ++j) {
             shz_xmtrx_init_fft_weights(a * j * ((float)size / (float)(len << 1)));
 
             // Butterfly operation for a particular Twiddle factor in a particular stage
-            for(int i = j; i < size; i += (len << 1)) {
+            for(size_t i = j; i < size; i += (len << 1)) {
                 shz_complex_t* x = &s[i];
                 shz_complex_t* y = &s[i + len];
 
@@ -47,8 +47,8 @@ void shz_fft(shz_complex_t* s, size_t size) {
         }
     }
 
-    int j = 0, k;
-    for(int i = 1; i < (size - 1); i++) {
+    size_t j = 0, k;
+    for(size_t i = 1; i < (size - 1); i++) {
         k = size / 2;
         
         while(k <= j) {
